List, object loading shouldn't need top-level [ or { block
Check if object text starts with { or list text starts with [ and add
the braces if necessary to the top-level construct if they aren't
already there. So you can do "Field: Value, Field2: Value2" and load it
as an object instead of having to put {} around that text. The {} is
required every time and in the exact same place if you're loading text
that represents an object, so why require it to be retyped every time?